Enhance error handling and update server port configuration

+++ backend/controllers/turfController.js
@@ -36,6 +36,10 @@ export const createProduct = async (req, res) => {
 export const deleteProduct =  async (req, res) => {
     const { id } = req.params
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid ID" });
+    }
+
     try {
         const deletedTurf = await Turf.findById(id);
         await Turf.findByIdAndDelete(id);
